Template Saving/Loading
I've added template saving/loading to effectively keep all your work on
thumbnails from tournament to tournament. Currently I am writing a
client side youtube uploading plugin for the software that will allow
videos to be rendered then uploaded. This will either be chosen by the
user to be uploaded automatically after rendering or for a later time.
This in which will be saved.