Added Rename and Fixes
Added the rename button so users are now able to rename the file if an error ocurred.
Users will also not be forced out of the program if they exit the Stream File window when specifying the FLV location.
Future:
Allowing personal team names, so no '&' will be placed into the teams slot. This will be happening by only inputting 1 member of the team, in this case the team name. If the other slot is to be empty, then the '&' will not be added.