Fixed Resize + Added New Feature
Fixed the resizing features for devices not operating at 1920x1080. Oddly enough the laptop I use is at this display but it still goofs up. No matter though, using the new feature fixes displays.

New Feature added allows teams to be entered with a single name. Example: Team Name -> Double Shine. If placed in either P1, P2 slot for team 1 without a name in the other, the program will recognize this and remove the '&' sign.
Same happens with P3 and P4 team.